Remove obsolete caps lock customizations
As of 01a7a5bd8a1e521756d1ec34769119ead5eee0b3, we're no longer using
`better-caps-lock.lua` to provide the ctrl/esc customizations for the
caps lock key. Therefore we can:

- Move the WindowLayout mappings out of `better-caps-lock.lua` and put
  them in `windows.lua`
- Move the Hammerspoon reload mapping out of `better-caps-lock.lua` and
  put it in `init.lua`
- Remove `better-caps-lock.lua`
