+++ hammerspoon/init.lua
@@ -10,5 +10,11 @@ require('control-escape')
 require('delete-words')
 require('hyper')
 require('super')
+require('windows')
 
 hs.notify.new({title='Hammerspoon', informativeText='Ready to rock 🤘'}):send()
+
+-- Use Control+` to reload Hammerspoon config
+hs.hotkey.bind({'ctrl'}, '`', nil, function()
+  hs.reload()
+end)
