+++ hammerspoon/windows.lua
@@ -166,3 +166,80 @@ function hs.window.nextScreen(win)
     win:moveToScreen(allScreens[1])
   end
 end
+
+--------------------------------------------------------------------------------
+-- Define WindowLayout Mode
+--
+-- WindowLayout Mode allows you to manage window layout using keyboard shortcuts
+-- that are on the home row, or very close to it. Use BetterCapsLock+w to turn
+-- on WindowLayout mode. Then, use any shortcut below to perform a window layout
+-- action. For example, to send the window left, press and release
+-- BetterCapsLock+w, and then press h.
+--
+--   h/j/k/l => send window to the left/bottom/top/right half of the screen
+--   i => send window to the upper left quarter of the screen
+--   o => send window to the upper right quarter of the screen
+--   , => send window to the lower left quarter of the screen
+--   . => send window to the lower right quarter of the screen
+--   return => make window full screen
+--   n => send window to the next monitor
+--------------------------------------------------------------------------------
+
+windowLayoutMode = hs.hotkey.modal.new({}, 'F16')
+
+-- Bind the given key to call the given function and exit WindowLayout mode
+function windowLayoutMode.bindWithAutomaticExit(mode, key, fn)
+  mode:bind({}, key, function()
+    mode:exit()
+    fn()
+  end)
+end
+
+windowLayoutMode:bindWithAutomaticExit('return', function()
+  hs.window.focusedWindow():maximize()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('space', function()
+  hs.window.focusedWindow():centerWithFullHeight()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('h', function()
+  hs.window.focusedWindow():left()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('j', function()
+  hs.window.focusedWindow():down()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('k', function()
+  hs.window.focusedWindow():up()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('l', function()
+  hs.window.focusedWindow():right()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('i', function()
+  hs.window.focusedWindow():upLeft()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('o', function()
+  hs.window.focusedWindow():upRight()
+end)
+
+windowLayoutMode:bindWithAutomaticExit(',', function()
+  hs.window.focusedWindow():downLeft()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('.', function()
+  hs.window.focusedWindow():downRight()
+end)
+
+windowLayoutMode:bindWithAutomaticExit('n', function()
+  hs.window.focusedWindow():nextScreen()
+end)
+
+-- Use Control+w to enter WindowLayout Mode
+hs.hotkey.bind({'ctrl'}, 'w', function()
+  windowLayoutMode:enter()
+end)
